pref: 拆分出功能与本体
