+++ ghv/Properties/launchSettings.json
@@ -1,0 +1,8 @@
+{
+  "profiles": {
+    "ghv": {
+      "commandName": "Project",
+      "commandLineArgs": "contribute"
+    }
+  }
+}
